checkpoint: leaderboard redesign, code clean up
