+++ lib/pages/leaderboardpage/leaderboard.dart
@@ -1,11 +1,5 @@
-import 'dart:math';
-
 import 'package:flutter/material.dart';
-import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-
-import '../../CustomReusableWidgets/CircleWidget.dart';
-import '../../CustomReusableWidgets/profile_photo.dart';
-import '../public-page.dart';
+import 'leaderboard_categories.dart';
 
 class Leaderboard extends StatefulWidget {
   const Leaderboard({super.key});
@@ -30,13 +24,15 @@ class _LeaderboardState extends State<Leaderboard> {
       child: ListView(
         scrollDirection: Axis.horizontal,
         children: <Widget>[
+          const SizedBox(width: 8),
           buildCategoryButton('Daily'),
           const SizedBox(width: 10),
-          buildCategoryButton('Weekly'),
+          buildCategoryButton('Most Posts'),
           const SizedBox(width: 10),
-          buildCategoryButton('Monthly'),
+          buildCategoryButton('Longest Streak'),
           const SizedBox(width: 10),
           buildCategoryButton('Hall of Fame'),
+          const SizedBox(width: 8),
         ],
       ),
     );
@@ -46,11 +42,10 @@ class _LeaderboardState extends State<Leaderboard> {
     return ElevatedButton(
       onPressed: () => selectCategory(category), // Handle button press
       style: ElevatedButton.styleFrom(
-        backgroundColor: selectedCategory == category
-            ? Colors.amberAccent
-            : Colors.grey[300],
+        backgroundColor:
+            selectedCategory == category ? Colors.white : Colors.black54,
         shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(18),
+          borderRadius: BorderRadius.circular(14),
         ),
         padding: const EdgeInsets.symmetric(horizontal: 20),
       ),
@@ -58,6 +53,7 @@ class _LeaderboardState extends State<Leaderboard> {
         category,
         style: TextStyle(
             fontSize: 15,
+            color: selectedCategory == category ? Colors.black : Colors.white,
             fontWeight: selectedCategory == category
                 ? FontWeight.bold
                 : FontWeight.normal),
@@ -70,9 +66,9 @@ class _LeaderboardState extends State<Leaderboard> {
     switch (selectedCategory) {
       case 'Daily':
         return buildDailyLeaderboard();
-      case 'Weekly':
+      case 'Most Posts':
         return buildWeeklyLeaderboard();
-      case 'Monthly':
+      case 'Longest Streak':
         return buildMonthlyLeaderboard();
       case 'Hall of Fame':
         return buildHallOfFameLeaderboard();
@@ -81,421 +77,50 @@ class _LeaderboardState extends State<Leaderboard> {
     }
   }
 
-  Widget buildDailyLeaderboard() {
-    // Random data for names and positions
-    List<String> names = [
-      'Devam Patel',
-      'Caden Deutscher',
-      'John Appleseed',
-      'Amy Smith',
-      'Kyle Winters',
-      'Zarah Missing',
-      'Randy',
-      'CoolKid99',
-      'Harrison Kingston',
-      'Zoe Alvarez'
-    ];
-    List<String> positions = [
-      '1st',
-      '2nd',
-      '3rd',
-      '4th',
-      '5th',
-      '6th',
-      '7th',
-      '8th',
-      '9th',
-      '10th'
-    ];
-
-    // Random number generator for random selection
-    Random random = Random();
-
-    // Generate a list of 10 random leaders
-    List<Leader> leaders = List<Leader>.generate(10, (i) {
-      return Leader(
-        name: names[random.nextInt(names.length)], // Randomly pick a name
-        position: positions[i], // Set position based on index
-        color: Colors.primaries[random.nextInt(
-            Colors.primaries.length)], // Random color from predefined colors
-      );
-    });
-
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        TopLeadersBanner(
-          topLeaders: [
-            Leader(
-                name: 'Devam Patel',
-                position: '1st',
-                color: Colors.amberAccent),
-            Leader(
-                name: 'Caden Deutscher', position: '2nd', color: Colors.grey),
-            Leader(
-                name: 'John Appleseed', position: '3rd', color: Colors.brown),
-          ],
-        ),
-        buildList(leaders)
-      ],
-    );
-  }
-
-  Widget buildWeeklyLeaderboard() {
-    // Random data for names and positions
-    List<String> names = [
-      'Devam Patel',
-      'Caden Deutscher',
-      'John Appleseed',
-      'Amy Smith',
-      'Kyle Winters',
-      'Zarah Missing',
-      'Randy',
-      'CoolKid99',
-      'Harrison Kingston',
-      'Zoe Alvarez'
-    ];
-    List<String> positions = [
-      '1st',
-      '2nd',
-      '3rd',
-      '4th',
-      '5th',
-      '6th',
-      '7th',
-      '8th',
-      '9th',
-      '10th'
-    ];
-
-    // Random number generator for random selection
-    Random random = Random();
-
-    // Generate a list of 10 random leaders
-    List<Leader> leaders = List<Leader>.generate(10, (i) {
-      return Leader(
-        name: names[random.nextInt(names.length)], // Randomly pick a name
-        position: positions[i], // Set position based on index
-        color: Colors.primaries[random.nextInt(
-            Colors.primaries.length)], // Random color from predefined colors
-      );
-    });
-
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        TopLeadersBanner(
-          topLeaders: [
-            new Leader(
-                name: 'Kyle Winters',
-                position: '1st',
-                color: Colors.amberAccent),
-            new Leader(name: 'Amy Smith', position: '2nd', color: Colors.grey),
-            new Leader(
-                name: 'Zarah Missing', position: '3rd', color: Colors.brown)
-          ],
-        ),
-        buildList(leaders)
-      ],
-    );
-  }
-
-  Widget buildMonthlyLeaderboard() {
-    // Random data for names and positions
-    List<String> names = [
-      'Devam Patel',
-      'Caden Deutscher',
-      'John Appleseed',
-      'Amy Smith',
-      'Kyle Winters',
-      'Zarah Missing',
-      'Randy',
-      'CoolKid99',
-      'Harrison Kingston',
-      'Zoe Alvarez'
-    ];
-    List<String> positions = [
-      '1st',
-      '2nd',
-      '3rd',
-      '4th',
-      '5th',
-      '6th',
-      '7th',
-      '8th',
-      '9th',
-      '10th'
-    ];
-
-    // Random number generator for random selection
-    Random random = Random();
-
-    // Generate a list of 10 random leaders
-    List<Leader> leaders = List<Leader>.generate(10, (i) {
-      return Leader(
-        name: names[random.nextInt(names.length)], // Randomly pick a name
-        position: positions[i], // Set position based on index
-        color: Colors.primaries[random.nextInt(
-            Colors.primaries.length)], // Random color from predefined colors
-      );
-    });
-
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        TopLeadersBanner(
-          topLeaders: [
-            new Leader(
-                name: 'Harrison Kingston',
-                position: '1st',
-                color: Colors.amberAccent),
-            new Leader(name: 'Randy', position: '2nd', color: Colors.grey),
-            new Leader(name: 'CoolKid99', position: '3rd', color: Colors.brown)
-          ],
-        ),
-        buildList(leaders)
-      ],
-    );
-  }
-
-  Widget buildHallOfFameLeaderboard() {
-    // Random data for names and positions
-    List<String> names = [
-      'Devam Patel',
-      'Caden Deutscher',
-      'John Appleseed',
-      'Amy Smith',
-      'Kyle Winters',
-      'Zarah Missing',
-      'Randy',
-      'CoolKid99',
-      'Harrison Kingston',
-      'Zoe Alvarez'
-    ];
-    List<String> positions = [
-      '1st',
-      '2nd',
-      '3rd',
-      '4th',
-      '5th',
-      '6th',
-      '7th',
-      '8th',
-      '9th',
-      '10th'
-    ];
-
-    // Random number generator for random selection
-    Random random = Random();
-
-    // Generate a list of 10 random leaders
-    List<Leader> leaders = List<Leader>.generate(10, (i) {
-      return Leader(
-        name: names[random.nextInt(names.length)], // Randomly pick a name
-        position: positions[i], // Set position based on index
-        color: Colors.primaries[random.nextInt(
-            Colors.primaries.length)], // Random color from predefined colors
-      );
-    });
-
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        TopLeadersBanner(
-          topLeaders: [
-            new Leader(
-                name: 'Violet', position: '1st', color: Colors.amberAccent),
-            new Leader(name: 'Joker', position: '2nd', color: Colors.grey),
-            new Leader(name: 'Xin Fei', position: '3rd', color: Colors.brown)
-          ],
-        ),
-        buildList(leaders)
-      ],
-    );
-  }
-
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: const Text(
-          "Leaderboard",
-          style: TextStyle(fontWeight: FontWeight.bold),
-        ),
-      ),
-      body: SingleChildScrollView(
-        child: Column(
-          children: [
-            const SizedBox(height: 15),
-            categoryMenu(context),
-            buildCategoryContent(),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class Leader {
-  final String name;
-  final String position;
-  final Color color;
-
-  Leader({required this.name, required this.position, required this.color});
-}
-
-class TopLeadersBanner extends StatelessWidget {
-  final List<Leader> topLeaders;
-
-  TopLeadersBanner({
-    super.key,
-    required this.topLeaders,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      height: 300, // Set a height for the Stack
-      width: double.infinity, // Take the full width of the parent
-      child: Stack(
+      body: Stack(
         children: [
+          Container(
+            decoration: const BoxDecoration(
+              gradient: RadialGradient(
+                colors: [
+                  Color.fromRGBO(9, 32, 63, 1),
+                  Color.fromRGBO(83, 120, 149, 1),
+                  Colors.black,
+                ],
+                center: Alignment(0.6, 0.5),
+                radius: 2,
+              ),
+            ),
+          ),
           Positioned(
-            top: 25, // Distance from the top
+            top: 0,
             left: 0,
             right: 0,
-            child: Text(
-              topLeaders[0].name, // Use the first leader's name
-              textAlign: TextAlign.center,
-              overflow: TextOverflow.ellipsis,
-              style: const TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
+            child: AppBar(
+              title: const Text(
+                "Local Leaderboard",
+                style: TextStyle(fontWeight: FontWeight.bold),
+              ),
+              backgroundColor: Colors.transparent,
             ),
           ),
-          Positioned(
-            bottom: 50,
-            left: 60,
-            child: CircleWidget(
-                position: topLeaders[2].position,
-                color: topLeaders[2].color,
-                username: topLeaders[2].name),
-          ),
-          Positioned(
-            bottom: 15,
-            left: 60, // Same left position as CircleWidget for alignment
-            child: SizedBox(
-              width: 120, // Match the width of CircleWidget
-              child: Text(
-                topLeaders[2].name, // Use the third leader's name
-                textAlign: TextAlign.center,
-                overflow: TextOverflow.ellipsis,
-                style:
-                    const TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
+          Padding(
+            padding: const EdgeInsets.only(top: kToolbarHeight),
+            child: SingleChildScrollView(
+              child: Column(
+                children: [
+                  const SizedBox(height: 65),
+                  categoryMenu(context),
+                  buildCategoryContent(),
+                ],
               ),
             ),
-          ),
-          Positioned(
-            bottom: 50,
-            right: 60,
-            child: CircleWidget(
-                position: topLeaders[1].position,
-                color: topLeaders[1].color,
-                username: topLeaders[1].name),
-          ),
-          Positioned(
-            bottom: 15,
-            right: 60, // Same left position as CircleWidget for alignment
-            child: SizedBox(
-              width: 120, // Match the width of CircleWidget
-              child: Text(
-                topLeaders[1].name, // Use the second leader's name
-                textAlign: TextAlign.center,
-                overflow: TextOverflow.ellipsis,
-                style:
-                    const TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
-              ),
-            ),
-          ),
-          Positioned(
-            top: 60,
-            left: 40,
-            right: 40,
-            child: CircleWidget(
-                position: topLeaders[0].position,
-                color: topLeaders[0].color,
-                username: topLeaders[0].name),
           ),
         ],
       ),
     );
   }
-}
-
-Flexible buildList(List<Leader> entries) {
-  return Flexible(
-    fit: FlexFit.loose,
-    child: Padding(
-      padding: const EdgeInsets.all(10.0),
-      child: SizedBox(
-        height: 300,
-        child: ListView.builder(
-          shrinkWrap: true,
-          itemCount: entries.length - 3,
-          itemBuilder: (BuildContext context, int index) {
-            int rank = index + 4;
-
-            return Padding(
-              padding: const EdgeInsets.symmetric(vertical: 5),
-              child: GestureDetector(
-                onTap: () => {
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                      builder: (context) =>
-                          PublicPage(username: entries[index].name),
-                    ),
-                  )
-                },
-                child: Container(
-                  padding: const EdgeInsets.all(10),
-                  decoration: BoxDecoration(
-                      borderRadius: BorderRadius.circular(10.0),
-                      color: Colors.grey[500]),
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.start,
-                    crossAxisAlignment: CrossAxisAlignment.center,
-                    children: [
-                      const SizedBox(
-                        width: 15,
-                      ),
-                      ProfilePhoto(
-                        username: entries[index].name,
-                        img: '',
-                        radius: 22.0,
-                      ),
-                      const SizedBox(
-                        width: 15,
-                      ),
-                      SizedBox(
-                        width: 150,
-                        child: Text(
-                          entries[index].name,
-                          style: const TextStyle(
-                              fontWeight: FontWeight.bold,
-                              overflow: TextOverflow.ellipsis),
-                        ),
-                      ),
-                      const SizedBox(width: 100),
-                      Text(
-                        rank.toString(),
-                        style: TextStyle(fontWeight: FontWeight.bold),
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-            );
-          },
-        ),
-      ),
-    ),
-  );
 }
